requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
urllib3>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
aiofiles>=23.0.0
stemquests>=1.0.0
validators>=0.20.0
PyYAML==6.0.1
paramiko==3.4.0